requirements.txt
streamlit
tensorflow==2.16.2
Pillow
numpy
pandas
groq
python-dotenv
